requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
tqdm
torch
torchvision
Pillow
opencv-python
logging
configparser